requirements.txt
beautifulsoup4>=4.14.2,
jupyter>=1.1.1,
matplotlib>=3.10.6,
mwparserfromhell>=0.7.2,
neo4j>=6.0.2,
networkx>=3.5,
numpy>=1.26.0,
pandas>=2.3.3,
python-louvain>=0.16,
requests>=2.32.5,
tqdm>=4.65.0,
google-generativeai>=0.8.0,
python-dotenv>=1.0.0,
